trip-journey-assistant/pyproject.toml
[project]
name = "journey-assistant"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a,otel-gcp]>=1.20.0",
    "litellm>=1.80.7",
    "mcp-weather-server>=0.6.0",
    "nacos-sdk-python>=3.0.2",
    "opentelemetry-distro[otlp]>=0.58b0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "tripsphere>=0.1.0",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["uv_build>=0.9.14,<0.10.0"]
build-backend = "uv_build"

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[tool.uv.sources]
tripsphere = { workspace = true }

[tool.uv.workspace]
members = ["libs/tripsphere"]

[tool.ruff]
target-version = "py312"
extend-exclude = ["libs/tripsphere"]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pyright]
typeCheckingMode = "strict"
include = ["src", "tests"]
